NO-SNOW: Rename exceptions to errors
This change alignes naming with old drivers
