+++ python/src/snowflake/connector/__init__.py
@@ -10,7 +10,7 @@ from typing import Any
 from ._internal.api_client.c_api import register_default_logger_callback
 from .connection import Connection
 from .cursor import Cursor
-from .exceptions import (
+from .errors import (
     DatabaseError,
     DataError,
     Error,
