+++ python/src/snowflake/connector/errors.py
